Pull in remote-unwinding hunks
A few patches related to more complete remote unwinding.  Note that
building the application now requires a custom compilation of libunwind
to be available--not because we need any particular modifications to the
library, but because the package manager doesn't ship with the
architecture-specifi remote unwinding static library needed for building
the profiler now.