Broken commit.  Add options handling
This is a broken commit, but pushing because I might go on paternity
leave any day now.  :)

Remaining work is to review line 167-168 of src/dd-prof.c, making usre
that execvp() is getting the right args.

Probably also need to make sure that the default values are populated
correctly in the DDRequest part of the context struct, as well as
hooking the `params` struct up to the profiling interface in the
callback.