Add debug symbols to build.
One of these days we need to have an adult conversation about this
makefile, because it's pretty sad.  Like not even sad in a "haha, this
is a hack and everybody understands that" way, but a "wow, this is just
sloppy work and it's a little embarrassing that you can continue to edit
this like it's no big deal" way.