Makefile: add build-time compiler detection
This is an embarrassing hack, but we're in a situation where e.g., the
default gcc is never the one we want, but we're not quite sure which one
we can get.  Same for clang.  Thanks Ubuntu!

(There are almost certainly silly implications--notably this means that the
user is incapable of setting a preference for the compiler version;
maybe it's time for a configure script?)