Add online help and help generation
The online help for ddprof is generated at time of execution, since we
have a tight binding between the commandline options, environment
variables, the help strings, and all of that kind of stuff.  Rather than
generating the code dynamically from config (in which case, we could
generate the help from there too), we keep config-as-code in a tight,
tabular format and generate the help at runtime.

This adds a script that generates a dumb skeleton .md in docs and just
passes the output of the ddprof help command into the file.  The user
still needs to actually run this step and commit the results and junk.