Fix regression in collatz
If you build collatz with too high an optimization setting, the compiler
will optimize-out some of the nested call structure, which is not what
we want.  Whereas we used to inherit the optimization flag from the
outer Makefile, here we just set it manually.  Maybe in the future we
should just drop the -O mode?