Refactor events interface, add perf_event override
Single-character events were getting in the way of providing a clean
syntax for overriding the event sampling rate.  These are provisional
names, but this way we can do something like `ddprof -e sCPU,1000` to
let the user override the sampling frequency.  One point of contention
we're going to have to deal with is the fact that 'events per second' is
a fairly transparent unit, but the dimensional analysis for 'seconds per
second' (sampling frequency for time-based events, such as wallclock
time) is maybe a little bit harder--people are used to representing
frequency, not sample counts, for time.  TBD.

The documentation for events is now generated programmatically at
runtime, so no more discrepancy between the help and the application.

The old collatz example isn't useful, we removed it.

Some of the events require different settings (usually following the
kernel), so a new system was added to set new modes to overriding the
defaults.