Ensure intermediate input params are on heap
The new X_FREE macro was intended to clear heap-allocated members of the
profiling context.  However, X_DFLT and X_CASE weren't updated to ensure
the defaulting or population code actually heap-allocated these
members--previously, some of these fields were referenced directly from
the environment, since they would be used throughout the lifetime of the
application without modification and it didn't make sense to allocate or
free them.  Now, doing so is a semantic requirement.

This defect arose as a segfault during initialization, before we install
the segfault pretty-printer.  There are a few other enabling factors.

 * In my review of the static analysis, I've historically ignored this
   section because it's been the source of nothing but false positives
   from both gcc and llvm
 * I disabled safety checks in CI while we're still pre-beta, since I
   didn't want to have to fix some offending parts of the build system
   yet
 * The last few releases were handled entirely in CI

This bumps the version to 0.4.3