Address deadlocks; add sendlast
Early-process was still susceptible to deadlocks.  In the last patch, I
tried to resolve this by adding a semaphore, but there is an additional
timing race I had neglected to consider in the first pass.  Rather than
adding even more complex handling, I removed the semaphore and created a
failure state in `getfd()` stemming from the emission of `sendfail()`,
which uses an inappropriate cmsg configuration to produce an error in
the recipient.  Adding that failure channel looks to have simplified the
code and cleared it up.

There were a few lingering reachable printf statements in libraries.
Added differential failure returns when relevant and pushed the
description up to the logging system shipped in ddprof.c

Added a new mode, sendlast, which will emit one final pprof if the
instrumented application is detected to have gone away.  This is rarely
useful,l since fast-exiting applications might close before unwinding
can symbolize.  Still, 60 seconds is a long time so maybe this helps.

Bumped version to 0.4.4