Support DD_PROFILING_ENABLED
This patch adds support for the DD_PROFILING_ENABLED environment
variable, as well as the DD_PROFILING_NATIVE_ENABLED override.  The
reason for both is, we don't currently provide a facility by which to
easily enable ddprof but disable an inner profiler (for example, if the
ddprof target is a Python script running the DataDog Python profiler) or
vice-versa.  Accordingly, DD_PROFILING_ENABLED will operate on both
inner and outer profilers in the absence of DD_PROFILING_NATIVE_ENABLED.
If DD_PROFILING_NATIVE_ENABLED is provided, it sets the behavior of the
native profiler without modifying inner profilers.

Note that these options are also available through the commandline.  I'm
not sure whether this is desirable, but for this first iteration I made
the commandline argument override the environment variable and set it to
the corresponding value for the inner environment.

In addition, this patch
 * Updates the documentation to reflect the new argument and document
   the old, but ignored, one
 * Provides new helper functions for detecting whether an arg is part of
   a given set (or its index in the set)
 * Slightly refactors arg processing to use the new helpers