Slightly improve runit.sh; perf.h safety
This is a mostly aesthetic improvement of the runit script--should make
it work better in environments where llvm-sanitizer is not available
(otherwise, it would fail silently previously).

Add a simple check to verify that the given file descriptor is valid in
`perfown`.  Technically, this check is implied by `mmap()`, but it can
be difficult to explain why the `mmap()` failed.  Maybe this is
irrelevant and everything can be inferred by errno, in which case that
check should be added instead of this one.