Fix defunct issues, split main into more helpers
It was reported that ddprof will persist as a zombie process if no
instrumentation could occur.  Although this isn't a serious problem by
itself, it could mean that there are resource leaks (imagine that a run
script is executed multiple times in a single container--those leaks add
up!)

It turns out that this was happening because ddprof forks twice and the
middle child exits immediately without properly cleaning up some of its
resources, such as pthread state and unix domain socket handles.  It's
likely that the process had undrained socket state (or futex), although
the exact nature of the leak was not explored.  Either way, I think we
fixed it.

This can be tested interactively by disabling perf events on a a debian-
or android- based base image with

```
sudo sh -c echo 3 > /proc/sys/kernel/perf_event_paranoid
```
(remember to reset it to 2 when you're done!)

and running the sleep test from runit.sh.  Before this change, you can
confirm that there are defunct processes.  After this change, there are
not.

I also went ahead and refactored `main()` slightly.  This was a plan for
a while, but I wanted to make sure the coordination primitives were more
fleshed out before committing to an API.  This still isn't ideal, since
there's a lot of weird coordination and global state, but it's better.
I would like to hold off on further substantial improvements until we've
added the coordination primitives into what is today the (send/get)fd
interface, since that would give us a clean IPC mechanism rather than a
thread barrier.