Simplify the startup sequence for ddprof
Thanks to an experiment from @r1viollet, we learned that
perf_event_paranoid==2 is sufficient for instrumenting other PIDs in the
same PID namespace.  This overturned previous experiments which had
shown that self-instrumentation is necessary.  I'm a little bit confused
by this result, but after testing it in multiple configurations it
appears to hold.  Accordingly, we can remove essentially all of the
coordination code during process startup.

This patch simplifies src/ddprof.c accordingly.  The only notable
difference is that we don't sequence the beginning of instrumentation as
tightly as we used to, so under very weird scheduler situations it may
be possible to miss early frames in the target process.  This should
generally not happen.