Remove bad unused function
The procutils ppid check had multiple calls to `getline()` with
unconsumed returns.  Failing to check one single getline is bad, but
failing to check multiple could segfault, if I understand the failure
mode correctly.

Since the function is unused and I don't want to fix it now, I went
ahead and rmoved it.