Resquence clear operations, add pprof to the mix
Also updates collrunner

In the prior work for fixing memory allocations, we did a last-minute
resequencing of where the free-type operations happened.  This was fine
and all, except it was only added to one of the terminal paths of
ddprof's main loop, not the main export loop.  This patch brings it back
into a once-per-export operation.