Erwan.viollet/use nobody user (#76)
* Update user id to nobody when performing mmap operation

* Change seteuid/setuid to setresuid

It appears that seteuid isn't sufficient for modifying the behavior of
mmap, and performing setuid with euid set to 0 makes the process
non-reversible.  At the same time, setuid() cannot be called after
seteuid() causes root permissions to be dropped.  We use the
Linux-specific setresuid() to effect both, allowing us to restore both
UIDs afterward.

* Add error trace on call to fctnl

Co-authored-by: sanchda <sanchda@users.noreply.github.com>