Remove float-type statistics
In a previous iteration of this WIP code, we had introduced a family of
functions for handling float-type statistics.  Superficially this is is
the correct thing to do, since dogstatsd admits float types and these
are often in use.

However, the ddprof codebase currently does not use this type and it
isn't clear we'll use it at all in the near-term.  Accordingly, I've
opted to remove the complexity of supporting a multi-type data model
without an exemplar of all types and simplified the code back to the
integral case.