Add richer diagnostics to stats/statsd interfaces
Previously, the error context of inner statsd functions was not passed
to the calling context, making it impossible to distinguish between
different errors.  In diagnosing a recent issue in the reliability
environment, I found it useful to surface slightly more granular
information, and I think customers would benefit from having logs with
this content as well.  This also restructures statsd and ddprof_stats
slightly to conform better with our standard error passing library.