Update stack size to use dyn_size
When reading off of the perf events ringbuffer, we were taking the size
of the stack to be the size of the data copied by the interface.
Instead, the interface provides a dyn_size quantity which should be used
for stack boundary computations.  As we only ever care about actually
reading off of the stack.