Fix IPC test path; add failover to statsd write
Just like IP-based UDP sockets, `write()` calls to an AF_UNIX,
SOCK_DGRAM can block if the kernel send buffer is full.  We're running
into this scenario in CI, which causes tests to block indefinitely.
Moreover, this could easily happen in customer infrastructure, which
would block the profiler indefinitely.

Rather than implementing any sort of failover, we just comment on the
error.  In full generality this may be a persistent error, so in a
followup commit I'll try to make sure this error is only emitted once
per export or something.