Update the statsd library to retry on EINTR
Whenever an application is interrupted by a signal, pending syscalls are
returned immediately with an EINTR.  The correct way to handle this is
via retry, but our statsd implementation did no such thing.

Arguably, it's perfectly fine to drop a single packet here and there,
but if retrying is easy we might as well just do that.  Now we do indeed
do that.