Fix race condition under contention in ringbuffer
All credit to @r1viollet for realizing that there was an issue and
essentially describing the conditions under which it arises.

The problem is we were marking the ringbuffer as available for
advancement even though we had yet to process the current sample.  When
the ringbuffer isn't full--no problem.  When it is full, the kernel will
over-write the current sample, probably with garbage.