Valgrind fixes: memory leaks (including removal of openssl) (#170)
- Drop OpenSSL for rustls
PR # 7 of libddprof proposes to use rustls. This fixes the issues raised by valgrind. It also reduces the number of bytes still reachable.

- Remove analyzis of unwinding errors
In debug mode we were using a special code that looked at proc maps. This code had leaks and some other issues.
As I do not want to redesign it, I am removing it for now.

- Tag leak fix
Free tags when allocated.

- Ring buffer : memory leak fix
The init / free sequence was asymmetrical causing a memory leak.

- Promote new dependency cache including rustls

- Adding libm as an expected dependency

 - Change worker exit flow to avoid leak
The exit flow was not freeing the ddprof_context. This change proposes to exit at the same place where the workers are spawned. This makes it easier to free the context when needed.