Clarify docs, scale timeout
Clarified the docs, since the behavior around agentless was
underspecified.

Added a fudge factor to make export timeouts scale more logically with
the size of the window.  With this implementation, the export paths will
hold for approximately a minute since the last export, taking into
account the size of the export window.  In some edge cases it might be
less than this and in some it might be more.  This mostly comes up when
testing at low values of -u under extremal network conditions.