Whole host profiling: avoid relying on file path (#194)
* Whole host profiling: avoid relying on file path
Use inode instead of path. Path information can be the same across several containers.
Also the path can be different with symlinks even if the underlying file is the same. The inode is a better source of truth.
It is not perfect, but in addition to size and offset information, it should be enough to guarantee uniqueness.

* Logger improvement (CPU usage)
Ensure logger does not evaluate arguments unless level is OK