Change the annotation strategy for registers
Since we're adding more registers, we have to now account for the fact
that x86 perf mask has a hole in it, and x86 registers are out-of-order
in general.  Plus, a lot of the stuff we were doing was somewhat
magical/mysterious, so fleshed out some more comments on why we're doing
things the way we're doing them.