Fix ringbuffer overflow
The size of the ringbuffer is given in bytes.  When that size is
initialized, we were using the *number* of registers, but not
multiplying it by the number of bytes per register.  Since the sample
object in the ringbuffer omits fields when they aren't specified by the
feature flags given to `perf_event_open()`, the ringbuffer elements were
always strictly smaller than our struct definition.  That struct
definition is part of the sizing criteria, so even though we were
undercounting the the register part of the sample, we overcounted in
other places to compensate.

Now that we're pulling in many more registers, the problem has become
noticeable to ASAN.  This fix just adds the size of the registers into
the computation.

Note that it would be preferable to link this back to the register size,
but it would be cumbersome to inject that quantity back into the header
deserializer.  Since we don't support any registers apart from 64-bit
ones, this is fine for now.