Fix static initialization order in allocation profiler
Compiler was generating a constructor for TrackerStaticState that is called before main and could be called after constructor of g_autostart that is starting allocation profiling.
This commit moves TrackerStaticState inside AllocationTracker as a non-static member. This adds an indirection to access this state, but overhead should be very low.
Also moves all free functions inside AllocationTracker, and exposes only the required functions.