Retry with kernel events
Proposal to save retry kernel logic at the level of the option.
Several options were considered.
  - Having several watchers
  - Having several configurations within one watcher
This version where we save the logic within the option is the simplest for now, but it has some drawbacks to extend the retry capability.
I chose this as I see extra work to support the retries at the level of hardware events.