Periodically check for newly loaded libraries
Rationale: some libraries (eg. nss libs) are dynamically loaded with __libc_dlopen_mode
and therefore are missed by dlopen hook.
Use a timer with a 59s interval with a signal handler that sets a boolean, this bool is
checked in all the hooks and triggers a check of loaded libraries.