Fix crash when using custom stack
When the user is using a custom stack (different from the thread's stack)
we would crash when copying the stack to the ring buffer.
This adds a bound check (checking that the current SP is within the stack).
If we are not within the thread's stack, we will just not consider this
sample. We would need acces to the runtime's API to understand the stack's
structure.
There is still a case where the stack could grow and we would not detect
the growth of the stack, however we would not crash, we would simply not
sample.