Report downstream pipeline failures to GH
It seems that since there are no "real" jobs in gitlab pipeline, only
a trigger for a downstream pipeline, gitsync does not push any status
check to github (status checks are associated to "real" jobs).
The workaround it to create an actual job whose status depends on
status of downstream pipeline.