Improve ELF segment matching
To compute elf bias, we need to match a Dso (ie. a mapping in process
address space) with the coresponding LOAD segment in elf file.
Before this change, we would use the Dso containing the address we want to
unwind, and make the assumption that an elf file contains only one LOAD
executable segment, and therefore that the matching segment would be the
unique executable LOAD segment of the elf file.
This assumption turns out to be wrong: an elf file can contain multiple
LOAD executable segments.
With this change, we use all the executable mappings of the elf file and
try to match them with LOAD segments by using the aligned offset and the
permissions of the mappings.
We consider only executable mappings LOAD segments (because we configure
perf in such a way that we receive only executable mmap events).
The matching is done by comparing mapping offset with aligned file offset
(file offset may not be aligned, in particular with lld, but
when segment is mmapped, mmap offset needs to be page aligned, file offset
is rounded down to satisfy segment p_align alignment) and memory
protection of segment.
Several segments can have the same aligned offsets and protections though,
thus making the matching ambiguous.
Any mapping can be used for matching, but some mappings might be missing
because of lost perf events. We start with the last mapping in address
order, because it is the one that has the most constraints: if we observed
n executable mappings, then we know that the last mapping is preceded by at
least n-1 other exec mappings and therefore cannot match the first n-1 LOAD
exec segments. If last mapping leads to an ambiguous match, we try again
with the second to last mapping and so on...