Allow anonymous mapping to truncate existing file mapping
Allow new mapping to truncate existing mappings if previous mapping is
a file and new mapping is anonymous because a LOAD segment can have
filesize < memsize as a size optimization when the end of segment is all
zeros (eg. bss). In that case elf loader / dlopen will map the whole
file + some extra space to cover the difference between memsize and file
size and then map an anonymous mapping over the part not in the file.