Add option to reorder events by timestamp across ring buffers (#387)
* Add option to reorder events by timestamp across ring buffers

Reorder events from ring buffers before processing them.
Events in each perf ring buffer are already ordered by timestamp. For MPSC ring buffers, there is no such guarantee.
The strategy is to dequeue events from ring buffers and push them in a priority queue, ordered by timestamp. And then process this priority queue.
When a ring buffer is empty, we cannot be sure that a new event with a timestamp less than a previously enqueued event will not be added to the ring buffer in the future.
That's why we arbitrarily define a maximum latency kMaxSampleLatency between the timestamp of an event and the time it appears in the ring buffer, and we process events with timestamps up to (now - kMaxSampleLatency).

For MPSC ring buffers, we dequeue events and push them into the priority queue until we reach an event with a timestamp greater than (now - kMaxSampleLatency) or the ring buffer is empty. Note that when an event with a timestamp greater than (now - kMaxSampleLatency) is dequeued, it is still pushed in the priority_queue.

For perf ring buffers, we ensure that at anytime at most one event from each ring buffer is in the priority queue. This ensures that events with identical timestamps in a ring buffer are processed in the same order as in
the ring buffer (priority queue is not stable, therefore pushing events with identical timestamps might permute them) and this also makes advancing the reader cursor position in the ring buffer easier since know that only one event has been read, we can just bump the reader cursor to the last read position.

When an event from a perf ring buffer is dequeued fron the priority queue, we advance the reader cursor position in the ring buffer to free the slot for the writer, and attempt to read the next event from the ring buffer if not empty.

When an event from a MPSC ring buffer is dequeued from the priority queue, we try to advance the reader cursor position in the ring buffer to free the slot for the writer. Since events might be out of order for this ring buffer, advancing is done by marking processed events as discarded and bumping the reader position until empty or we reach the first non-discarded event.

* Disable cpu affinity workarounds when event reordering is enabled
* Add a backpopulate metric