Allocation tracker - fix deadlock in log
When entering error paths in the allocation profiler, we would log.
This could create deadlocks while trying to acquire a lock to print timestamps in logs.
Here is the backtrace where the deadlock was found:
Thread 1 (Thread 0x7f0dffe31180 (LWP 1773386)):

This commit moves the logs to the debug level to make sure they would not print by default.
We can still trigger the logs in debug mode.